+++ Packages/com.unity.inputsystem/InputSystem/Controls/InputControl.cs
@@ -996,6 +996,9 @@ namespace UnityEngine.InputSystem
             }
         }
 
+        // Allows determining if the device has an associated state pointer
+        internal unsafe bool hasState => currentStatePtr != null;
+
         // This data is initialized by InputDeviceBuilder.
         internal InternedString m_Name;
         internal string m_Path;